+++ integer/src/mul/toom_3.rs
@@ -137,7 +137,7 @@ pub fn add_signed_mul_same_len(
     {
         let t1_short = &mut t1[..2 * n3];
         let overflow = mul::add_signed_mul_same_len(t1_short, Positive, a0, b0, &mut memory);
-        assert!(overflow == 0);
+        debug_assert!(overflow == 0);
         carry_c0 += add::add_signed_same_len_in_place(&mut c[..2 * n3], sign, t1_short);
         carry_c2 += add::add_signed_in_place(&mut c[2 * n3..4 * n3 + 2], -sign, t1_short);
         t1[2 * n3] = mul::mul_word_in_place(t1_short, 3);
@@ -157,7 +157,7 @@ pub fn add_signed_mul_same_len(
         b_eval[n3] = mul::add_mul_word_same_len_in_place(&mut b_eval[..n3], 2, b1);
         b_eval[n3] += mul::add_mul_word_in_place(&mut b_eval[..n3], 4, b2);
         let overflow = mul::add_signed_mul_same_len(t1, Positive, a_eval, b_eval, &mut memory);
-        assert!(overflow == 0);
+        debug_assert!(overflow == 0);
     }
 
     // Evaluate at inf.
@@ -170,13 +170,13 @@ pub fn add_signed_mul_same_len(
         let (c_eval, mut memory) = memory.allocate_slice_fill(2 * n3 + 2, 0);
         let c_eval_short = &mut c_eval[..2 * n3_short];
         let overflow = mul::add_signed_mul_same_len(c_eval_short, Positive, a2, b2, &mut memory);
-        assert!(overflow == 0);
+        debug_assert!(overflow == 0);
         carry_c2 += add::add_signed_in_place(&mut c[2 * n3..4 * n3 + 2], -sign, c_eval_short);
         carry += add::add_signed_same_len_in_place(&mut c[4 * n3..], sign, c_eval_short);
         c_eval[2 * n3_short] = mul::mul_word_in_place(c_eval_short, 12);
         let overflow = add::sub_in_place(t1, &c_eval[..2 * n3_short + 1]);
         // 3V(0) + V(2) - 12V(inf) is never negative
-        assert!(!overflow);
+        debug_assert!(!overflow);
     }
 
     // Sign of V(-1).
@@ -203,7 +203,7 @@ pub fn add_signed_mul_same_len(
         b_eval[n3] += Word::from(add::add_same_len_in_place(&mut b_eval[..n3], b1));
 
         let overflow = mul::add_signed_mul_same_len(t2, Positive, a_eval, b_eval, &mut memory);
-        assert!(overflow == 0);
+        debug_assert!(overflow == 0);
         carry_c1 += add::add_signed_in_place(&mut c[n3..3 * n3 + 2], sign, t2);
 
         // Evaluate at -1.
@@ -222,17 +222,17 @@ pub fn add_signed_mul_same_len(
     }
     let (c_eval, mut memory) = memory.allocate_slice_fill(2 * (n3 + 1), 0);
     let overflow = mul::add_signed_mul_same_len(c_eval, Positive, a_eval, b_eval, &mut memory);
-    assert!(overflow == 0);
+    debug_assert!(overflow == 0);
     let overflow = add::add_signed_same_len_in_place(t2, value_neg1_sign, c_eval);
-    assert!(overflow == 0);
+    debug_assert!(overflow == 0);
     match value_neg1_sign {
         Positive => {
             let overflow = mul::add_mul_word_same_len_in_place(t1, 2, c_eval);
-            assert!(overflow == 0);
+            debug_assert!(overflow == 0);
         }
         Negative => {
             let overflow = mul::sub_mul_word_same_len_in_place(t1, 2, c_eval);
-            assert!(overflow == 0);
+            debug_assert!(overflow == 0);
         }
     }
 
@@ -260,6 +260,6 @@ pub fn add_signed_mul_same_len(
     carry_c3 += add::add_signed_word_in_place(&mut c[4 * n3 + 2..5 * n3 + 2], carry_c2);
     carry += add::add_signed_word_in_place(&mut c[5 * n3 + 2..], carry_c3);
 
-    assert!(carry.abs() <= 1);
+    debug_assert!(carry.abs() <= 1);
     carry
 }